Some updates

+++ FingerprintAnalysis/src/beans/TemplateTextBean.java
@@ -12,9 +12,11 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.Serializable;
+import java.rmi.server.RMIClassLoader;
 import java.util.Scanner;
 
 import javax.imageio.ImageIO;
+import javax.swing.JOptionPane;
 import javax.swing.JTextArea;
 
 import imageProcessing.ImageConvertor;
@@ -35,32 +37,39 @@ public class TemplateTextBean extends JTextArea implements Serializable {
 		setEditable(true);
 		setText("");
 
-		ImageConvertor imageConvertor = new ImageProcessing();
-		imageConvertor.setPathInputImage(pathImage);
-		BufferedImage bufferedImage = imageConvertor.convertToGrayscale();
-		File file = new File(resultFileNameWithoutExtension + ".png");
-		ImageIO.write(bufferedImage, "png", file);
-
-		String executableMindtctString = "./exe\\mindtct.exe";
-		Terminal.executeCommand(executableMindtctString + " -m1 " + resultFileNameWithoutExtension + ".png" + " " + resultFileNameWithoutExtension);
-		Terminal.executeCommand("rm *brw *dm *hcm *lcm *lfm *min *qm");
-		FileConvertor.convertIt(resultFileNameWithoutExtension + ".png", resultFileNameWithoutExtension);
-		if ((new File("trash//" + resultFileNameWithoutExtension + ".txt")).exists()) {
-			BufferedReader reader = new BufferedReader(new FileReader("trash//" + resultFileNameWithoutExtension + ".txt"));
-			String line = reader.readLine();
-			while (line != null) {
-				setText(getText() + line + "\n");
-				line = reader.readLine();
-			}
-			reader.close();
+		File file = new File(pathImage);
+		File fileBis = file;
+		if (ImageIO.read(file).getColorModel().getPixelSize() != 8) {
+			ImageConvertor imageConvertor = new ImageProcessing();
+			imageConvertor.setPathInputImage(file.getAbsolutePath());
+			BufferedImage convertedImage = imageConvertor.convertToGrayscale();
+			fileBis = new File("trash\\" + getFileNameWithoutExtension(file) + ".png");
+			ImageIO.write(convertedImage, "png", fileBis);
 		}
-		drawMinutaes(resultFileNameWithoutExtension, imageBean);
 
+		if(ImageIO.read(file).getColorModel().getPixelSize() == 1)
+	        JOptionPane.showMessageDialog(null, "Do not use binary images!", "Error", JOptionPane.INFORMATION_MESSAGE);
+		else {
+			String executableMindtctString = "./resources/exe\\mindtct.exe";
+			Terminal.executeCommand(executableMindtctString + " -m1 " + fileBis.getAbsolutePath() + " " + resultFileNameWithoutExtension);
+			Terminal.executeCommand("rm *brw *dm *hcm *lcm *lfm *min *qm");
+			FileConvertor.convertIt(fileBis.getAbsolutePath(), resultFileNameWithoutExtension);
+			if ((new File("trash//" + resultFileNameWithoutExtension + ".txt")).exists()) {
+				BufferedReader reader = new BufferedReader(new FileReader("trash//" + resultFileNameWithoutExtension + ".txt"));
+				String line = reader.readLine();
+				while (line != null) {
+					setText(getText() + line + "\n");
+					line = reader.readLine();
+				}
+				reader.close();
+			}
+			drawMinutaes(resultFileNameWithoutExtension, imageBean);
+		}
 		setEditable(false);
 	}
 
 	public static void clearRubbish(String resultFileNameWithoutExtension) {
-		Terminal.executeCommand("rm trash\\" + resultFileNameWithoutExtension + ".txt " + resultFileNameWithoutExtension + ".xyt " + resultFileNameWithoutExtension + ".png");
+		Terminal.executeCommand("rm " + resultFileNameWithoutExtension + ".xyt trash//*");
 	}
 
 	private void drawMinutaes(String resultFileNameWithoutExtension, ImageBean imageBean) throws Exception {
@@ -80,6 +89,12 @@ public class TemplateTextBean extends JTextArea implements Serializable {
 			}
 			scanner.close();
 		}
+	}
+	
+	private String getFileNameWithoutExtension(File file) throws Exception {
+		String fileName = "";
+		if (file != null && file.exists()) fileName = file.getName().replaceFirst("[.][^.]+$", "");
+		return fileName;
 	}
 
 }
